Moved vecmem::device_address_space out of details/
Since the main declaration of vecmem::device_atomic_ref
needs that enumeration as well.
