+++ core/include/vecmem/memory/device_address_space.hpp
